Remove FullReplAPI and ReplApiImpl
These are now unnecessary, since some stuff were moved out of ReplAPI.
